pyproject.toml
[tool.poetry]
name = "pierogis-live"
version = "0.1.0"
description = ""
authors = ["Kyle Moore <moore.kyle39@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-dotenv = "^0.15.0"
Flask = "^1.1.2"
ffmpeg-python = "^0.2.0"
Flask-Cors = "^3.0.9"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"
py-postgresql = "^1.2.2"

[tool.poetry.dev-dependencies]
ffmpeg-python = "^0.2.0"
Flask = "^1.1.2"
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
